+++ src/NzbDrone.Core/Books/Services/AddBookService.cs
@@ -119,11 +119,11 @@ namespace NzbDrone.Core.Books
             }
             catch (BookNotFoundException)
             {
-                _logger.Error("Book with Foreign Id {0} was not found, it may have been removed from Goodreads.", newBook.ForeignBookId);
+                _logger.Error("Book with foreign ID {0} was not found by the metadata provider.", newBook.ForeignBookId);
 
                 throw new ValidationException(new List<ValidationFailure>
                                               {
-                                                  new ValidationFailure("GoodreadsId", "A book with this ID was not found", newBook.ForeignBookId)
+                                                  new ValidationFailure("ForeignBookId", "A book with this ID was not found", newBook.ForeignBookId)
                                               });
             }