+++ src/Readarr.Api.V1/Config/DevelopmentConfigController.cs
@@ -1,10 +1,7 @@
 using System.Linq;
 using System.Reflection;
-using FluentValidation;
 using Microsoft.AspNetCore.Mvc;
-using NzbDrone.Common.Extensions;
 using NzbDrone.Core.Configuration;
-using NzbDrone.Core.Validation;
 using NzbDrone.Http.REST.Attributes;
 using Readarr.Http;
 using Readarr.Http.REST;
@@ -22,8 +19,6 @@ namespace Prowlarr.Api.V1.Config
         {
             _configFileProvider = configFileProvider;
             _configService = configService;
-
-            SharedValidator.RuleFor(c => c.MetadataSource).IsValidUrl().When(c => !c.MetadataSource.IsNullOrWhiteSpace());
         }
 
         protected override DevelopmentConfigResource GetResourceById(int id)
@@ -34,7 +29,7 @@ namespace Prowlarr.Api.V1.Config
         [HttpGet]
         public DevelopmentConfigResource GetDevelopmentConfig()
         {
-            var resource = DevelopmentConfigResourceMapper.ToResource(_configFileProvider, _configService);
+            var resource = DevelopmentConfigResourceMapper.ToResource(_configFileProvider);
             resource.Id = 1;
 
             return resource;