+++ src/Readarr.Api.V1/Config/DevelopmentConfigResource.cs
@@ -5,7 +5,6 @@ namespace Prowlarr.Api.V1.Config
 {
     public class DevelopmentConfigResource : RestResource
     {
-        public string MetadataSource { get; set; }
         public string ConsoleLogLevel { get; set; }
         public bool LogSql { get; set; }
         public int LogRotate { get; set; }
@@ -13,11 +12,10 @@ namespace Prowlarr.Api.V1.Config
 
     public static class DevelopmentConfigResourceMapper
     {
-        public static DevelopmentConfigResource ToResource(this IConfigFileProvider model, IConfigService configService)
+        public static DevelopmentConfigResource ToResource(this IConfigFileProvider model)
         {
             return new DevelopmentConfigResource
             {
-                MetadataSource = configService.MetadataSource,
                 ConsoleLogLevel = model.ConsoleLogLevel,
                 LogSql = model.LogSql,
                 LogRotate = model.LogRotate,